refactor: decompose session, substitution-sheet, achievements, share-card (batch 5)
