+++ __tests__/lib/strava.test.ts
@@ -22,10 +22,10 @@ const settingsSrc = [
   fs.readFileSync(path.resolve(__dirname, "../../components/settings/IntegrationsCard.tsx"), "utf-8"),
 ].join("\n");
 
-const sessionSrc = fs.readFileSync(
-  path.resolve(__dirname, "../../app/session/[id].tsx"),
-  "utf-8"
-);
+const sessionSrc = [
+  fs.readFileSync(path.resolve(__dirname, "../../app/session/[id].tsx"), "utf-8"),
+  fs.readFileSync(path.resolve(__dirname, "../../hooks/useSessionActions.ts"), "utf-8"),
+].join("\n");
 
 const layoutSrc = fs.readFileSync(
   path.resolve(__dirname, "../../app/_layout.tsx"),
