+++ components/ShareCard.tsx
@@ -1,10 +1,12 @@
 import React from "react";
-import { Platform, StyleSheet, View } from "react-native";
+import { StyleSheet, View } from "react-native";
 import { Text } from "@/components/ui/text";
 import MaterialCommunityIcons from "@expo/vector-icons/MaterialCommunityIcons";
 import { spacing } from "../constants/design-tokens";
 import { useThemeColors } from "@/hooks/useThemeColors";
 import { useColorScheme } from "@/hooks/useColorScheme";
+import { ShareCardStats } from "./share/ShareCardStats";
+import { ShareCardExercises } from "./share/ShareCardExercises";
 
 export type ShareCardExercise = {
   name: string;
@@ -30,7 +32,6 @@ export type ShareCardProps = {
   exercises: ShareCardExercise[];
 };
 
-const MAX_EXERCISES = 6;
 const CARD_WIDTH = 1080;
 
 function StarRow({ rating }: { rating: number }) {
@@ -54,9 +55,6 @@ export default function ShareCard(props: ShareCardProps) {
   const isDark = useColorScheme() === "dark";
   const { name, date, duration, sets, volume, unit, rating, prs, exercises } =
     props;
-
-  const displayExercises = exercises.slice(0, MAX_EXERCISES);
-  const remaining = exercises.length - MAX_EXERCISES;
 
   return (
     <View
@@ -100,46 +98,12 @@ export default function ShareCard(props: ShareCardProps) {
         </Text>
       </View>
 
-      {/* Stats row */}
-      <View
-        style={[
-          cardStyles.statsContainer,
-          { backgroundColor: colors.surfaceVariant },
-        ]}
-      >
-        <View style={cardStyles.statItem}>
-          <Text style={[cardStyles.statValue, { color: colors.onSurface }]}>
-            {duration}
-          </Text>
-          <Text
-            style={[cardStyles.statLabel, { color: colors.onSurfaceVariant }]}
-          >
-            Duration
-          </Text>
-        </View>
-        <View style={[cardStyles.statDivider, { backgroundColor: colors.outline }]} />
-        <View style={cardStyles.statItem}>
-          <Text style={[cardStyles.statValue, { color: colors.onSurface }]}>
-            {sets}
-          </Text>
-          <Text
-            style={[cardStyles.statLabel, { color: colors.onSurfaceVariant }]}
-          >
-            Sets
-          </Text>
-        </View>
-        <View style={[cardStyles.statDivider, { backgroundColor: colors.outline }]} />
-        <View style={cardStyles.statItem}>
-          <Text style={[cardStyles.statValue, { color: colors.onSurface }]}>
-            {volume}
-          </Text>
-          <Text
-            style={[cardStyles.statLabel, { color: colors.onSurfaceVariant }]}
-          >
-            Volume ({unit})
-          </Text>
-        </View>
-      </View>
+      <ShareCardStats
+        duration={duration}
+        sets={sets}
+        volume={volume}
+        unit={unit}
+      />
 
       {/* Rating */}
       {rating != null && rating >= 1 && (
@@ -148,85 +112,7 @@ export default function ShareCard(props: ShareCardProps) {
         </View>
       )}
 
-      {/* PRs */}
-      {prs.length > 0 && (
-        <View
-          style={[
-            cardStyles.prSection,
-            { backgroundColor: colors.primaryContainer },
-          ]}
-        >
-          <View style={cardStyles.prHeader}>
-            <Text style={[cardStyles.prTitle, { color: colors.onPrimaryContainer }]}>
-              🏆 New PRs
-            </Text>
-          </View>
-          {prs.map((pr, i) => (
-            <View key={i} style={cardStyles.prRow}>
-              <Text
-                style={[cardStyles.prName, { color: colors.onPrimaryContainer }]}
-                numberOfLines={1}
-                ellipsizeMode="tail"
-              >
-                {pr.name}
-              </Text>
-              <Text
-                style={[cardStyles.prValue, { color: colors.onPrimaryContainer }]}
-              >
-                {pr.value}
-              </Text>
-            </View>
-          ))}
-        </View>
-      )}
-
-      {/* Exercises */}
-      {displayExercises.length > 0 && (
-        <View style={cardStyles.exerciseSection}>
-          <Text
-            style={[
-              cardStyles.exerciseSectionTitle,
-              { color: colors.onSurfaceVariant },
-            ]}
-          >
-            Exercises
-          </Text>
-          {displayExercises.map((ex, i) => (
-            <View key={i} style={cardStyles.exerciseRow}>
-              <Text
-                style={[
-                  cardStyles.exerciseName,
-                  { color: colors.onSurface },
-                ]}
-                numberOfLines={1}
-                ellipsizeMode="tail"
-              >
-                {ex.name}
-              </Text>
-              <Text
-                style={[
-                  cardStyles.exerciseDetail,
-                  { color: colors.onSurfaceVariant },
-                ]}
-              >
-                {ex.weight
-                  ? `${ex.sets}×${ex.reps} @ ${ex.weight}`
-                  : `${ex.sets}×${ex.reps}`}
-              </Text>
-            </View>
-          ))}
-          {remaining > 0 && (
-            <Text
-              style={[
-                cardStyles.moreText,
-                { color: colors.onSurfaceVariant },
-              ]}
-            >
-              and {remaining} more
-            </Text>
-          )}
-        </View>
-      )}
+      <ShareCardExercises exercises={exercises} prs={prs} />
 
       {/* Footer */}
       <View
@@ -275,36 +161,6 @@ const cardStyles = StyleSheet.create({
     fontSize: 18,
     lineHeight: 24,
   },
-  statsContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    borderRadius: 16,
-    paddingVertical: spacing.lg,
-    paddingHorizontal: spacing.xl,
-    marginBottom: spacing.lg,
-  },
-  statItem: {
-    flex: 1,
-    alignItems: "center",
-  },
-  statValue: {
-    fontSize: 24,
-    fontWeight: "700",
-    ...Platform.select({
-      ios: { fontVariant: ["tabular-nums"] },
-      android: { fontVariant: ["tabular-nums"] },
-      default: {},
-    }),
-  },
-  statLabel: {
-    fontSize: 13,
-    marginTop: 2,
-  },
-  statDivider: {
-    width: 1,
-    height: 32,
-    marginHorizontal: spacing.sm,
-  },
   ratingSection: {
     alignItems: "center",
     marginBottom: spacing.lg,
@@ -312,65 +168,6 @@ const cardStyles = StyleSheet.create({
   starRow: {
     flexDirection: "row",
     gap: spacing.xs,
-  },
-  prSection: {
-    borderRadius: 16,
-    padding: spacing.lg,
-    marginBottom: spacing.lg,
-  },
-  prHeader: {
-    marginBottom: spacing.sm,
-  },
-  prTitle: {
-    fontSize: 18,
-    fontWeight: "700",
-  },
-  prRow: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    paddingVertical: spacing.xs,
-  },
-  prName: {
-    fontSize: 16,
-    fontWeight: "500",
-    flex: 1,
-    marginRight: spacing.sm,
-  },
-  prValue: {
-    fontSize: 16,
-    fontWeight: "700",
-  },
-  exerciseSection: {
-    marginBottom: spacing.lg,
-  },
-  exerciseSectionTitle: {
-    fontSize: 14,
-    fontWeight: "600",
-    textTransform: "uppercase",
-    letterSpacing: 1,
-    marginBottom: spacing.sm,
-  },
-  exerciseRow: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    paddingVertical: spacing.xs + 2,
-  },
-  exerciseName: {
-    fontSize: 16,
-    fontWeight: "500",
-    flex: 1,
-    marginRight: spacing.sm,
-  },
-  exerciseDetail: {
-    fontSize: 14,
-    fontWeight: "500",
-  },
-  moreText: {
-    fontSize: 14,
-    fontStyle: "italic",
-    marginTop: spacing.xs,
   },
   footer: {
     borderTopWidth: 1,
