+++ components/SubstitutionSheet.tsx
@@ -1,19 +1,19 @@
 import React, { useCallback, useMemo, useState } from "react";
-import { Alert, Platform, Pressable, StyleSheet, View } from "react-native";
+import { Alert, Platform, StyleSheet, View } from "react-native";
 import { Text } from "@/components/ui/text";
-import { Chip } from "@/components/ui/chip";
 import BottomSheet, {
   BottomSheetBackdrop,
   BottomSheetFlatList,
 } from "@gorhom/bottom-sheet";
 import type { Exercise, Equipment } from "../lib/types";
-import { EQUIPMENT_LABELS, DIFFICULTY_LABELS, MUSCLE_LABELS } from "../lib/types";
+import { MUSCLE_LABELS } from "../lib/types";
 import {
   findSubstitutions,
   type SubstitutionScore,
 } from "../lib/exercise-substitutions";
-import { radii } from "../constants/design-tokens";
 import { useThemeColors } from "@/hooks/useThemeColors";
+import { SubstitutionItem } from "./substitution/SubstitutionItem";
+import { EquipmentFilter } from "./substitution/EquipmentFilter";
 
 type Props = {
   sheetRef: React.RefObject<BottomSheet | null>;
@@ -22,106 +22,6 @@ type Props = {
   onSelect: (exercise: Exercise) => void;
   onDismiss: () => void;
 };
-
-function matchColor(score: number): string {
-  if (score >= 80) return "#2e7d32";
-  if (score >= 60) return "#f57f17";
-  return "#c62828";
-}
-
-function matchBgColor(score: number): string {
-  if (score >= 80) return "#e8f5e9";
-  if (score >= 60) return "#fff8e1";
-  return "#ffebee";
-}
-
-function SubstitutionItem({
-  item,
-  onPress,
-}: {
-  item: SubstitutionScore;
-  onPress: (exercise: Exercise) => void;
-}) {
-  const colors = useThemeColors();
-  const ex = item.exercise;
-
-  return (
-    <Pressable
-      style={[styles.item, { backgroundColor: colors.surface }]}
-      onPress={() => onPress(ex)}
-      accessibilityRole="button"
-      accessibilityLabel={`${ex.name}, ${item.score}% match, ${EQUIPMENT_LABELS[ex.equipment]}, ${DIFFICULTY_LABELS[ex.difficulty]}`}
-    >
-      <View style={styles.itemHeader}>
-        <Text
-          variant="body"
-          numberOfLines={1}
-          style={[styles.itemName, { color: colors.onSurface, fontWeight: "600" }]}
-        >
-          {ex.name}
-        </Text>
-        <View
-          style={[
-            styles.matchBadge,
-            { backgroundColor: matchBgColor(item.score) },
-          ]}
-        >
-          <Text variant="caption" style={[styles.matchText, { color: matchColor(item.score) }]}>
-            {item.score}%
-          </Text>
-        </View>
-      </View>
-      <View style={styles.itemMeta}>
-        <View
-          style={[
-            styles.equipBadge,
-            { backgroundColor: colors.surfaceVariant },
-          ]}
-        >
-          <Text variant="caption"
-            style={[styles.equipText, { color: colors.onSurfaceVariant }]}
-          >
-            {EQUIPMENT_LABELS[ex.equipment]}
-          </Text>
-        </View>
-        <View
-          style={[
-            styles.equipBadge,
-            { backgroundColor: colors.surfaceVariant },
-          ]}
-        >
-          <Text variant="caption"
-            style={[styles.equipText, { color: colors.onSurfaceVariant }]}
-          >
-            {DIFFICULTY_LABELS[ex.difficulty]}
-          </Text>
-        </View>
-      </View>
-      {ex.primary_muscles.length > 0 && (
-        <View style={styles.muscleRow}>
-          {ex.primary_muscles.map((m) => (
-            <View
-              key={m}
-              style={[
-                styles.muscleChip,
-                { backgroundColor: colors.secondaryContainer },
-              ]}
-            >
-              <Text variant="caption"
-                style={[
-                  styles.muscleText,
-                  { color: colors.onSecondaryContainer },
-                ]}
-              >
-                {MUSCLE_LABELS[m]}
-              </Text>
-            </View>
-          ))}
-        </View>
-      )}
-    </Pressable>
-  );
-}
 
 export default function SubstitutionSheet({
   sheetRef,
@@ -265,33 +165,11 @@ export default function SubstitutionSheet({
             </View>
           ) : (
             <>
-              {availableEquipment.length > 1 && (
-                <View style={styles.filterRow} accessibilityRole="toolbar" accessibilityLabel="Equipment filter">
-                  <Chip
-                    selected={equipmentFilter === null}
-                    onPress={() => setEquipmentFilter(null)}
-                    compact
-                    style={styles.filterChip}
-                    accessibilityLabel="Show all equipment"
-                  >
-                    All
-                  </Chip>
-                  {availableEquipment.map((eq) => (
-                    <Chip
-                      key={eq}
-                      selected={equipmentFilter === eq}
-                      onPress={() =>
-                        setEquipmentFilter(equipmentFilter === eq ? null : eq)
-                      }
-                      compact
-                      style={styles.filterChip}
-                      accessibilityLabel={`Filter by ${EQUIPMENT_LABELS[eq]}`}
-                    >
-                      {EQUIPMENT_LABELS[eq]}
-                    </Chip>
-                  ))}
-                </View>
-              )}
+              <EquipmentFilter
+                availableEquipment={availableEquipment}
+                equipmentFilter={equipmentFilter}
+                onFilterChange={setEquipmentFilter}
+              />
 
               {scored.length === 0 ? (
                 <View style={styles.emptyState}>
@@ -351,53 +229,8 @@ const styles = StyleSheet.create({
   muscleText: {
     lineHeight: 16,
   },
-  filterRow: {
-    flexDirection: "row",
-    flexWrap: "wrap",
-    gap: 6,
-    marginBottom: 12,
-  },
-  filterChip: {
-    marginBottom: 2,
-  },
   listContent: {
     paddingBottom: 32,
-  },
-  item: {
-    padding: 12,
-    borderRadius: radii.md,
-    marginBottom: 8,
-  },
-  itemHeader: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-    marginBottom: 4,
-  },
-  itemName: {
-    flex: 1,
-    fontWeight: "600",
-    marginRight: 8,
-  },
-  matchBadge: {
-    paddingHorizontal: 8,
-    paddingVertical: 2,
-    borderRadius: 12,
-  },
-  matchText: {
-    fontWeight: "700",
-  },
-  itemMeta: {
-    flexDirection: "row",
-    gap: 6,
-    marginBottom: 4,
-  },
-  equipBadge: {
-    paddingHorizontal: 6,
-    paddingVertical: 2,
-    borderRadius: 8,
-  },
-  equipText: {
   },
   emptyState: {
     paddingVertical: 48,
